+++ chemtools/__init__.py
@@ -24,7 +24,7 @@
 """The Main ChemTools Package."""
 
 
-from chemtools.wrappers import *
+from chemtools.wrappers2 import *
 from chemtools.toolbox import *
 from chemtools.conceptual import *
 from chemtools.denstools import *
