+++ chemtools/toolbox/kinetic.py
@@ -27,7 +27,7 @@
 import numpy as np
 
 from chemtools.utils.utils import doc_inherit
-from chemtools.wrappers.molecule import Molecule
+from chemtools.wrappers2.molecule import Molecule
 from chemtools.denstools.densbased import DensGradTool, DensGradLapTool, DensGradLapKedTool
