+++ chemtools/toolbox/test/test_conceptual_condensed.py
@@ -28,8 +28,8 @@ import numpy as np
 
 from numpy.testing import assert_raises, assert_equal, assert_almost_equal
 
-from chemtools.wrappers.molecule import Molecule
-from chemtools.wrappers.grid import MolecularGrid
+from chemtools.wrappers2.molecule import Molecule
+from chemtools.wrappers2.grid import MolecularGrid
 from chemtools.toolbox.conceptual import CondensedConceptualDFT
 try:
     from importlib_resources import path
