+++ chemtools/toolbox/test/test_conceptual_global.py
@@ -29,7 +29,7 @@ import numpy as np
 from numpy.testing import assert_raises, assert_almost_equal
 
 from chemtools.toolbox.conceptual import GlobalConceptualDFT
-from chemtools.wrappers.molecule import Molecule
+from chemtools.wrappers2.molecule import Molecule
 try:
     from importlib_resources import path
 except ImportError:
