+++ chemtools/toolbox/test/test_oxidation.py
@@ -28,9 +28,9 @@ import numpy as np
 from numpy.testing import assert_equal, assert_almost_equal, assert_approx_equal
 
 from horton import ProAtomDB
-from chemtools.wrappers.grid import MolecularGrid
-from chemtools.wrappers.molecule import Molecule
-from chemtools.wrappers.part import DensPart
+from chemtools.wrappers2.grid import MolecularGrid
+from chemtools.wrappers2.molecule import Molecule
+from chemtools.wrappers2.part import DensPart
 from chemtools.toolbox.oxidation import EOS
