+++ chemtools/toolbox/topology.py
@@ -26,7 +26,7 @@
 
 import numpy as np
 
-from chemtools.wrappers.molecule import Molecule
+from chemtools.wrappers2.molecule import Molecule
 from chemtools.utils.cube import UniformGrid
 from chemtools.topology.critical import Topology
 from chemtools.outputs.vmd import print_vmd_script_topology
