+++ chemtools/wrappers2/part.py
@@ -28,8 +28,8 @@ import numpy as np
 from horton import ProAtomDB
 from horton.scripts.wpart import wpart_schemes
 
-from chemtools.wrappers.molecule import Molecule
-from chemtools.wrappers.grid import MolecularGrid
+from chemtools.wrappers2.molecule import Molecule
+from chemtools.wrappers2.grid import MolecularGrid
 
 
 __all__ = ['DensPart']
